Moved the controller methods outside of the Student group class and into the Controller class.
